Restore tests/test_lib.py with shared FileObjectImageInfo per reviewer feedback.
