+++ tests/img_info.py
@@ -2,52 +2,11 @@
 """Tests for Img_Info."""
 
 import os
-import threading
 import unittest
 
 import pytsk3
 
-
-class FileObjectImageInfo(pytsk3.Img_Info):
-  """Img_Info that uses a file-like object.
-
-  Thread-safety: pytsk3 may invoke read() concurrently from multiple
-  threads (true under free-threaded Python; possible even with the
-  GIL via cooperative thread switches across the seek+read pair).
-  Most file-like objects implement read positioning as a stateful
-  seek+read on a single fd, so concurrent calls would race. We
-  serialize the seek/read pair under a per-instance lock so that
-  pytsk3 callers can share one FileObjectImageInfo across threads.
-  """
-
-  def __init__(
-      self, file_object, file_size, image_type=pytsk3.TSK_IMG_TYPE_RAW):
-    """Initializes the image object."""
-    if not file_object:
-      raise ValueError(u'Missing file-like object.')
-
-    self._file_object = file_object
-    self._file_size = file_size
-    self._read_lock = threading.Lock()
-    pytsk3.Img_Info.__init__(self, url='', type=image_type)
-
-  def close(self):
-    """Closes the volume IO object."""
-    with self._read_lock:
-      self._file_object = None
-
-  def read(self, offset, size):
-    """Reads a byte string from the image object at the specified offset."""
-    with self._read_lock:
-      file_object = self._file_object
-      if file_object is None:
-        return b''
-      file_object.seek(offset, os.SEEK_SET)
-      return file_object.read(size)
-
-  def get_size(self):
-    """Retrieves the size."""
-    return self._file_size
+import test_lib
 
 
 class TSKImgInfoTestCase(unittest.TestCase):
@@ -144,19 +103,19 @@ class TSKImgInfoFileObjectTest(TSKImgInfoTestCase):
 
   def testInitialize(self):
     """Test the initialize functionality."""
-    img_info = FileObjectImageInfo(self._file_object, self._file_size)
+    img_info = test_lib.FileObjectImageInfo(self._file_object, self._file_size)
     self._testInitialize(img_info)
     img_info.close()
 
   def testGetSize(self):
     """Test the get size functionality."""
-    img_info = FileObjectImageInfo(self._file_object, self._file_size)
+    img_info = test_lib.FileObjectImageInfo(self._file_object, self._file_size)
     self._testGetSize(img_info)
     img_info.close()
 
   def testRead(self):
     """Test the read functionality."""
-    img_info = FileObjectImageInfo(self._file_object, self._file_size)
+    img_info = test_lib.FileObjectImageInfo(self._file_object, self._file_size)
     self._testRead(img_info)
     img_info.close()
 
@@ -175,7 +134,7 @@ class TSKImgInfoFileObjectWithDetectTest(TSKImgInfoTestCase):
 
   def testInitialize(self):
     """Test the initialize functionality."""
-    img_info = FileObjectImageInfo(
+    img_info = test_lib.FileObjectImageInfo(
         self._file_object, self._file_size,
         image_type=pytsk3.TSK_IMG_TYPE_DETECT)
     self._testInitialize(img_info)
@@ -183,7 +142,7 @@ class TSKImgInfoFileObjectWithDetectTest(TSKImgInfoTestCase):
 
   def testGetSize(self):
     """Test the get size functionality."""
-    img_info = FileObjectImageInfo(
+    img_info = test_lib.FileObjectImageInfo(
         self._file_object, self._file_size,
         image_type=pytsk3.TSK_IMG_TYPE_DETECT)
     self._testGetSize(img_info)
@@ -191,7 +150,7 @@ class TSKImgInfoFileObjectWithDetectTest(TSKImgInfoTestCase):
 
   def testRead(self):
     """Test the read functionality."""
-    img_info = FileObjectImageInfo(
+    img_info = test_lib.FileObjectImageInfo(
         self._file_object, self._file_size,
         image_type=pytsk3.TSK_IMG_TYPE_DETECT)
     self._testRead(img_info)
@@ -210,19 +169,19 @@ class TSKImgInfoFileObjectLargeSizeTest(TSKImgInfoTestCase):
 
   def testInitialize(self):
     """Test the initialize functionality."""
-    img_info = FileObjectImageInfo(self._file_object, self._file_size)
+    img_info = test_lib.FileObjectImageInfo(self._file_object, self._file_size)
     self._testInitialize(img_info)
     img_info.close()
 
   def testGetSize(self):
     """Test the get size functionality."""
-    img_info = FileObjectImageInfo(self._file_object, self._file_size)
+    img_info = test_lib.FileObjectImageInfo(self._file_object, self._file_size)
     self._testGetSize(img_info)
     img_info.close()
 
   def testRead(self):
     """Test the read functionality."""
-    img_info = FileObjectImageInfo(self._file_object, self._file_size)
+    img_info = test_lib.FileObjectImageInfo(self._file_object, self._file_size)
     self._testRead(img_info)
     img_info.close()
